+++ docs/changes.rst
@@ -5,6 +5,7 @@ Not yet released.
 
 * Fixed crash on certain component filemasks.
 * Improved accurancy of the consecutive duplicated words check.
+* Added support for Pagure pull requests.
 
 Weblate 4.3.1
 -------------